refactor(select): MenuItem -> Option
