+++ packages/select/src/components/Option/Component.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import cn from 'classnames';
-import { ItemShape, MenuItemProps } from '../../Component';
+import { ItemShape, OptionProps } from '../../Component';
 
 import styles from './index.module.css';
 
-export const MenuItem = <T extends ItemShape>({
+export const Option = <T extends ItemShape>({
     item,
     selected,
     highlighted,
     itemRenderer,
-}: MenuItemProps<T>) => {
+}: OptionProps<T>) => {
     return (
         <div
-            className={cn(styles.item, {
+            className={cn(styles.option, {
                 [styles.highlighted]: highlighted,
                 [styles.selected]: selected,
             })}
