+++ packages/select/src/components/Option/index.module.css
@@ -1,6 +1,6 @@
 @import '../../../../themes/default.css';
 
-.item {
+.option {
     @mixin system_14-18_regular;
 
     display: block;
